Fix build Clang: -Wsign-conversion su size_t/iterator (visto in CI)
GCC non abilita -Wsign-conversion dentro -Wconversion di default, Clang si.
find_start_code() ora prende size_t* invece di int* per la lunghezza dello
start code (3 o 4, mai negativa), e un piccolo helper accentra il cast a
difference_type richiesto da vector::iterator + size_t.

+++ annexb.h
@@ -46,7 +46,7 @@ public:
 
 private:
     std::vector<uint8_t> buf_;
-    static size_t find_start_code(const std::vector<uint8_t>& b, size_t from, int* len);
+    static size_t find_start_code(const std::vector<uint8_t>& b, size_t from, size_t* len);
 };
 
 // Annex-B: start code a 4 byte + NAL "nudo".
